% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/misc_py.R
\name{source_lines}
\alias{source_lines}
\title{Source specific lines in an R file}
\usage{
source_lines(file, lines)

source_lines(file, lines)
}
\arguments{
\item{file}{character string with the path to the file to source.}

\item{lines}{numeric vector of lines to source in \code{file}.}
}
\description{
Source specific lines in an R file

Source specific lines in an R file
}
